requirements.txt
datasets
scipy
numpy
transformers
OpenAI
evaluate
sacrebleu
pycocoevalcap
jieba
mecab-python3
unidic-lite